refactor: Inject embedded resources after CSC
