+++ src/Directory.Build.props
@@ -73,19 +73,25 @@
     -->
 	  <UnoUIUseRoslynSourceGenerators Condition="'$(TargetFramework)'=='net461'">true</UnoUIUseRoslynSourceGenerators>
 
-	  <!--
+	<!--
     Force the compilation using Roslyn Generators on monoandroid
     -->
 	  <UnoUIUseRoslynSourceGenerators Condition="'$(TargetFramework)'=='monoandroid11.0' or '$(TargetFramework)'=='monoandroid12.0' or '$(TargetFramework)'=='xamarinmac20'">true</UnoUIUseRoslynSourceGenerators>
-	  <!--
+	<!--
     Force the compilation using Roslyn Generators on xamarinios
     -->
-	  <UnoUIUseRoslynSourceGenerators Condition="'$(TargetFramework)'=='xamarinios10'">true</UnoUIUseRoslynSourceGenerators>
+	<UnoUIUseRoslynSourceGenerators Condition="'$(TargetFramework)'=='xamarinios10'">true</UnoUIUseRoslynSourceGenerators>
 
 	  <!--
     Force the compilation using Roslyn Generators for Skia targets
     -->
     <UnoUIUseRoslynSourceGenerators Condition="'$(MSBuildProjectName)'!='' and $(MSBuildProjectName.EndsWith('Skia'))">true</UnoUIUseRoslynSourceGenerators>
+
+	<!--
+    Force the compilation using Roslyn Generators for Wasm targets
+    -->
+
+	<UnoUIUseRoslynSourceGenerators Condition="'$(MSBuildProjectName)'!='' and $(MSBuildProjectName.EndsWith('Wasm'))">true</UnoUIUseRoslynSourceGenerators>
 </PropertyGroup>
 
   <!--
