+++ src/Directory.Build.targets
@@ -50,8 +50,8 @@
 	<PackageReference Update="Uno.Core.Extensions.Disposables" Version="4.0.1" />
 	<PackageReference Update="Uno.Core.Extensions.Compatibility" Version="4.0.1" />
 	<PackageReference Update="Uno.Diagnostics.Eventing" Version="2.0.1" />
-	<PackageReference Update="Uno.Wasm.Bootstrap" Version="2.1.0" />
-	<PackageReference Update="Uno.Wasm.Bootstrap.DevServer" Version="2.1.0" />
+	<PackageReference Update="Uno.Wasm.Bootstrap" Version="4.0.0-dev.191" />
+	<PackageReference Update="Uno.Wasm.Bootstrap.DevServer" Version="4.0.0-dev.191" />
 	<PackageReference Update="xamarin.build.download" Version="0.11.3" PrivateAssets="all" />
 	<PackageReference Update="MSTest.TestAdapter" Version="2.2.10" />
 	<PackageReference Update="MSTest.TestFramework" Version="2.2.8" />
